requirements.txt
python-dotenv
pyserial
meshcore
influxdb3-python
requests